Finalized separation of eatiers and Campus Information sections
