+++ app/src/main/java/mits/uwi/com/ourmobileenvironment/HTTP_RequestHandlers/RestaurantErrorListener.java
@@ -10,9 +10,7 @@ import java.util.List;
 
 import mits.uwi.com.ourmobileenvironment.EateriesActivity;
 import mits.uwi.com.ourmobileenvironment.R;
-import mits.uwi.com.ourmobileenvironment.Transport.TransportFragment;
 import mits.uwi.com.ourmobileenvironment.adapters.EateriesAdapter;
-import mits.uwi.com.ourmobileenvironment.campusinformationfragments.EateriesFragment;
 import mits.uwi.com.ourmobileenvironment.campusinformationfragments.Restaurant;
 
 /**
@@ -24,16 +22,7 @@ public class RestaurantErrorListener implements Response.ErrorListener {
     private Class<Restaurant> restaurantClass;
     private Toast internetConnection;
     private ArrayList<Restaurant> restaurants;
-    private EateriesFragment eateriesFragment;
     private EateriesActivity eateriesActivity;
-
-
-    public RestaurantErrorListener(Class<Restaurant> restaurantClass,Context mCtx,EateriesFragment eateriesFragment,ArrayList<Restaurant>restaurants){
-        this.restaurantClass=restaurantClass;
-        this.eateriesFragment=eateriesFragment;
-        internetConnection=Toast.makeText(mCtx,"Please Check Internet Connection",Toast.LENGTH_SHORT);
-        this.restaurants=restaurants;
-    }
 
     public RestaurantErrorListener(Class<Restaurant> restaurantClass, Context mCtx,
                                    EateriesActivity eateriesActivity,
@@ -47,17 +36,11 @@ public class RestaurantErrorListener implements Response.ErrorListener {
 
     @Override
     public void onErrorResponse(VolleyError error){
-        EateriesFragment.EateriesAdapter adapter;
-        EateriesAdapter newAdapter = null;
+        EateriesAdapter adapter;
 
         List<Restaurant> objlist = Restaurant.listAll(restaurantClass);
 
-        try{
-            adapter = eateriesFragment.getAdap();
-        } catch (NullPointerException e) {
-            adapter = null;
-            newAdapter = eateriesActivity.getAdapter();
-        }
+       adapter = eateriesActivity.getAdapter();
 
 
         if (objlist.isEmpty()){
@@ -67,22 +50,13 @@ public class RestaurantErrorListener implements Response.ErrorListener {
             restaurants.clear();
             for (Restaurant transport:objlist){
                 restaurants.add(transport);
-                if (adapter != null){
-                    adapter.Add(transport);
-                }
-                else if (newAdapter != null){
-                    newAdapter.Add(transport);
-                }
+                adapter.Add(transport);
+
             }
 
-            if (adapter != null){
-                eateriesFragment.refreshView();
-                eateriesFragment.getActivity().findViewById(R.id.progress_bar).setVisibility(View.GONE);
-            }
-            else if (newAdapter != null){
-                eateriesActivity.refreshView();
-                eateriesActivity.findViewById(R.id.progress_bar).setVisibility(View.GONE);
-            }
+            eateriesActivity.refreshView();
+            eateriesActivity.findViewById(R.id.progress_bar).setVisibility(View.GONE);
+
         }
     }
 }
