+++ app/src/main/java/mits/uwi/com/ourmobileenvironment/HTTP_RequestHandlers/RestaurantListener.java
@@ -17,9 +17,7 @@ import java.util.ArrayList;
 
 import mits.uwi.com.ourmobileenvironment.EateriesActivity;
 import mits.uwi.com.ourmobileenvironment.R;
-import mits.uwi.com.ourmobileenvironment.Transport.TransportFragment;
 import mits.uwi.com.ourmobileenvironment.adapters.EateriesAdapter;
-import mits.uwi.com.ourmobileenvironment.campusinformationfragments.EateriesFragment;
 import mits.uwi.com.ourmobileenvironment.campusinformationfragments.Restaurant;
 
 /**
@@ -32,24 +30,9 @@ public class RestaurantListener implements Response.Listener<JSONObject> {
     private Restaurant currentRes;
     private ArrayList resList;
     private Class<Restaurant> restaurantClass;
-    private EateriesFragment eateriesFragment;
     private EateriesActivity eateriesActivity;
     private ImageLoader imageLoader;
 
-
-
-    public RestaurantListener(String listname, ArrayList<Restaurant> resList,
-                              EateriesFragment eateriesFragment,Context resctx,
-                              Class<Restaurant> restaurantClass){
-        this.listname=listname;
-        this.eateriesFragment=eateriesFragment;
-        this.internalError= Toast.makeText(
-                resctx,
-                "Internal error occured please restart application",
-                Toast.LENGTH_SHORT);
-        this.resList=resList;
-        this.restaurantClass=restaurantClass;
-    }
 
     public RestaurantListener(String listname, ArrayList<Restaurant> resList,
                               EateriesActivity eateriesActivity, Context resctx,
@@ -69,14 +52,9 @@ public class RestaurantListener implements Response.Listener<JSONObject> {
     public void onResponse(JSONObject response){
         Gson gson = new Gson();
         String status;
-        EateriesFragment.EateriesAdapter adapter;
-        EateriesAdapter newAdapter = null;
-        try {
-            adapter = eateriesFragment.getAdap();
-        }catch (NullPointerException e){
-            adapter = null;
-            newAdapter = eateriesActivity.getAdapter();
-        }
+        EateriesAdapter adapter;
+
+        adapter = eateriesActivity.getAdapter();
         try {
             restuarantList=response.getJSONArray(listname);
             status=response.getString("Status");
@@ -86,12 +64,8 @@ public class RestaurantListener implements Response.Listener<JSONObject> {
             for (int i=0; i<restuarantList.length();i++){
                 currentRes =gson.fromJson(restuarantList.getJSONObject(i).toString(),
                                           restaurantClass );
-                if (adapter != null){
-                    adapter.Add(currentRes);
-                }
-                else if (newAdapter != null ){
-                    newAdapter.Add(currentRes);
-                }
+                adapter.Add(currentRes);
+
                 if (status.equals("200")){
                     currentRes.save();
                 }
@@ -103,16 +77,10 @@ public class RestaurantListener implements Response.Listener<JSONObject> {
                 }
             }
 
-            if (adapter != null){
-                eateriesFragment.refreshView();
-                eateriesFragment.getActivity()
-                        .findViewById(R.id.progress_bar)
-                        .setVisibility(View.GONE);
-            }
-            else if(newAdapter != null){
-                eateriesActivity.refreshView();
-                eateriesActivity.findViewById(R.id.progress_bar).setVisibility(View.GONE);
-            }
+
+            eateriesActivity.refreshView();
+            eateriesActivity.findViewById(R.id.progress_bar).setVisibility(View.GONE);
+
         }
         catch (JSONException e){
             Log.d("Restaurant exception",e.toString());
