+++ app/src/main/res/layout/activity_eateries.xml
@@ -1,8 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent" android:layout_height="match_parent"
     android:padding="0dp"
     >
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/eateries_toolbar"
+        android:layout_height="@dimen/abc_action_bar_default_height_material"
+        android:layout_width="match_parent"
+        android:minHeight="@dimen/abc_action_bar_default_height_material"
+        android:background="#000000"/>
+
+
     <ProgressBar
         android:id="@+id/progress_bar"
         style="?android:progressBarStyle"
@@ -17,9 +26,7 @@
         android:layout_height="match_parent"
         android:layout_width="match_parent"
         android:id="@+id/rv"
-        >
-
-    </android.support.v7.widget.RecyclerView>
+        />
 
 
 </LinearLayout>
