+++ app/src/main/res/values/strings.xml
@@ -1,13 +1,14 @@
 <resources>
-    <!--<string name="app_name">OurMobileEnvironment</string>-->
     <string name="app_name">UWI MONA</string>
     <string name="action_settings">Settings</string>
 
     <!-- Landing strings -->
     <string name="navigation_drawer_open">Open</string>
     <string name="navigation_drawer_close">Close</string>
+
     <!-- View Article Chooser Strings -->
-    <string name="view_article_string">View this news article with</string>
+
+    <!-- Fab  Labels -->
     <string name='campusinfo_fab'>Campus Information</string>
     <string name='transportation_fab'>Transportation</string>
     <string name='ourvle_fab'>OurVLE</string>
@@ -163,6 +164,15 @@
 
 
     <string name="title_activity_campus_information">Campus Information</string>
+    <string-array name="campus_info_snippet_titles">
+        <item>We are fully Accredited</item>
+        <item>We are the home of the Pelicans</item>
+        <item>We are a gree, progressive campus</item>
+        <item>Who we serve</item>
+        <item>Mission Statement</item>
+        <item>The UWI\'s Vision</item>
+    </string-array>
+
     <string name="campus_info_snippet_title1">We are fully Accredited</string>
     <string name="campus_info_snippet_title2">We are the home of the Pelicans</string>
     <string name="campus_info_snippet_title3">We are a green, progressive campus</string>
@@ -176,6 +186,51 @@
         development designed to support the social and economic growth of the Caribbean region, the
         provision of community service, and the enrichment of the academic experience.
     </string>
+
+    <string-array name="campus_info_snippet_body">
+        <item>
+            UWI Mona is a fully accredited tertiary-level institution.
+            Our Programmes - available across our 5 faculties,
+            8 schools and 12 specialist centres - have
+            been closely examined and evaluated and we
+            are proud of the \'flying colors\' we
+            have continuously received
+        </item>
+        <item>
+            We are the home of the Pelicans,
+            our enduring symbol of the humble Caribbean
+            bird that sits atop our coat of
+            arms highlighting the attitude, and altitude of
+            the distinctive UWI student.
+        </item>
+        <item>
+            UWI is increasingly modeling eco friendly sustainable
+            practices in its policy, operations and research.
+        </item>
+        <item>
+            Within the region, the UWI Mona serves 16
+            countries of the Commonwealth Caribbean\nThe UWI: One
+            Institution, Four Campuses\nUWI Mona: One Campus,
+            Two Locations\nKingston Campus, Mona | Western
+            Jamaica Campus, Montego Bay
+        </item>
+        <item>
+            Our Mission Statement reflects the primary purpose
+            of our University, that is, the reason
+            for its existence.To advance education and
+            create knowledge through excellence in teaching, research,
+            innovation, public service, intellectual leadership and outreach
+            in order to support the inclusive (social,
+            economic, political, cultural, environmental) development of the
+            Caribbean region and beyond.
+        </item>
+        <item>
+            By 2017, the University will be globally
+            recognised as a regionally integrated, innovative, internationally
+            competitive university, deeply rooted in all aspects of Caribbean
+            development and committed to serving the diverse people of the region and beyond.
+        </item>
+    </string-array>
 
     <string name="campus_info_snippet_body1">
         UWI Mona is a fully accredited tertiary-level institution. Our Programmes - available across
@@ -236,6 +291,8 @@
     <string name="action_res">Restaurants</string>
     <string name="action_deli">Delis</string>
     <string name="action_fav">Favourites</string>
+    <string name="title_activity_eateries">Eateries</string>
+
     <string name="title_activity_bus_search">BusSearchActivity</string>
 
 
@@ -341,12 +398,12 @@
 
     <string name="title_activity_news_view">News</string>
 
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+
+
     <string name="title_activity_campus_information_sub">CampusInformationSubActivity</string>
 
-    <string name="hello_world">Hello world!</string>
-    <string name="title_activity_eateries">EateriesActivity</string>
+
+
 
 
 </resources>
